backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
pydantic
python-dotenv
httpx
neuro-san